add arrows icons
