--- frontend/src/components/DatePicker/DatePicker.scss
@@ -1,4 +1,0 @@
-.primary{
-width: 100px !important;
-background-color: aqua !important;
-}
